+++ CHANGELOG.md
@@ -4,6 +4,7 @@
  * Add support for custom headers at client and request level
  * Add support for request
  * Update README adapters section
+ * Bug #6 - allow halon to handle resource list responses
 
 ### 0.0.1
 Initial release